Sprint 3

+++ src/main/java/com/transsurf/pe/repositorio/DestinoRepositorio.java
@@ -4,7 +4,7 @@ import com.transsurf.pe.entidades.Ciudad;
 import com.transsurf.pe.entidades.Destino;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DestinoRepositorio extends JpaRepository<Destino, Integer> {
+public interface DestinoRepositorio extends JpaRepository<Destino, Long> {
 
     public void deleteByCiudadIs(Ciudad ciudad);
